% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dates.R
\name{clean_dates}
\alias{clean_dates}
\title{Clean messy date variables}
\usage{
clean_dates(x, quietly = FALSE)
}
\arguments{
\item{x}{a vector of dates to be cleaned}

\item{quietly}{a TRUE/FALSE of whether to show warnings for dropped observations
(default is FALSE, so warnings are shown)}
}
\description{
Clean messy date variables
}
\details{
Any observation with only text will be set to missing (NA).
Properly spelt text names of months will be translated from French and Portuguese to English.
If there is a date and year along with the month name this will be changed to
an appropriate date (nb. if there is an ending to date e.g. \emph{1st}, \emph{2nd} etc. this
will not work).

Dates imported from Excel are identified by containing 5 numbers and starting
with the number 4. This will only work for dates after July 2009 and before
November 2036.(This could be updated to attempt all those containing any 5
numbers).

Anything else in a date formate will be kept; regardless of the separator
(dots, slashes or dashes).

All observations are then passed through \code{\link[lubridate:parse_date_time]{lubridate::parse_date_time()}} to
produce appropriate dates.
}
\author{
Alice Carr, Alex Spina
}
