% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_M_and_E.R
\name{clean_kpi}
\alias{clean_kpi}
\title{Clean country monitoring and evaluation excel tool data (key performance indicators)}
\usage{
clean_kpi(inputfile)
}
\arguments{
\item{inputfile}{a data frame produced by the \code{\link[covidmonitor:merge_kpi]{covidmonitor::merge_kpi()}}}
}
\description{
Clean country monitoring and evaluation excel tool data (key performance indicators)
}
\author{
Alice Carr, Alex Spina
}
