% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_M_and_E.R
\name{merge_kpi}
\alias{merge_kpi}
\title{Merge country monitoring and evaluation excel tool data (key performance indicators)}
\usage{
merge_kpi(
  inputdirectory,
  isotomerge = "AFRO",
  outputdirectory = tempdir(),
  outputname = "kpi_merged_",
  wide = TRUE,
  clean = TRUE,
  variabledict = system.file("extdata", "mne_dictionary.xlsx", package =
    "covidmonitor"),
  cleaningdict = system.file("inst", "cleaning_dictionary.xlsx", package =
    "covidmonitor")
)
}
\arguments{
\item{inputdirectory}{path to folder containing excel tool datasets}

\item{isotomerge}{3 letter ISO-code for country of interest, in quotation marks,
combine multiple countries with c(...), default is "AFRO" to read in all available.}

\item{outputdirectory}{path to folder where merged file is to be saved (must exist already)}

\item{outputname}{character string to name merged file}

\item{wide}{logical (TRUE/FALSE) of whether the output data frame should be
in wide format (default), else will produce long format data}

\item{clean}{logical (TRUE/FALSE) of whether to clean the output data frame}

\item{variabledict}{path to dictionary file containing variable definitions.
Default is a predefined within the package. For details of how to use your own
file see \href{https://github.com/R4IDSR/covidmonitor}{README}}

\item{cleaningdict}{path to dictionary file containing cleaning rule definitions
for country names.
Default is a predefined within the package. For details of how to use your own
file see \href{https://github.com/R4IDSR/covidmonitor}{README}}
}
\description{
Merge country monitoring and evaluation excel tool data (key performance indicators)
}
\author{
Alice Carr, Alex Spina
}
