% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelist_merge.R
\name{merge_linelist}
\alias{merge_linelist}
\title{Merge COVID-19 linelists from WHO AFRO}
\usage{
merge_linelist(
  inputdirectory,
  outputdirectory = tempdir(),
  outputname = "Merged_linelist_",
  isotomerge = "AFRO",
  quietly = TRUE,
  variabledict = system.file("extdata", "linelist_dictionary.xlsx", package =
    "covidmonitor")
)
}
\arguments{
\item{inputdirectory}{path to folder containing datasets}

\item{outputdirectory}{path to folder where merged file is to be saved (must exist already)}

\item{outputname}{character string to name merged file}

\item{isotomerge}{3 letter ISO-code for country of interest, in quotation marks,
combine multiple countries with c(...), default is "AFRO" to read in all available.
See details for countries currently supported.}

\item{quietly}{TRUE/FALSE of whether to show message of which country being
processed as the function runs through files. If false will show message for
each country (default = TRUE).}

\item{variabledict}{path to dictionary file containing variable definitions.
Default is a predefined within the package. For details of how to use your own
file see \href{https://github.com/R4IDSR/covidmonitor}{README}}
}
\description{
Merge COVID-19 linelists from WHO AFRO
}
\details{
Countries currently supported include:
"BFA", "CIV", "COD", "COG", "DZA", "GIN", "KEN", "LBR", "MOZ",
"MUS", "NAM", "NER", "RWA", "SEN", "SLE", "STP", "SWZ", "SYC",
"TCD", "UGA", "ZWE"
}
\author{
Alice Carr, Alex Spina
}
