% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_M_and_E.R
\name{merge_M_and_E}
\alias{merge_M_and_E}
\title{Merge weekly report monitoring and evaluation data.
Currently based on Guinea template; to be further in case of non-templated
documents}
\usage{
merge_M_and_E(
  inputdirectory,
  outputdirectory = tempdir(),
  outputname = "filename",
  template = TRUE,
  wide = TRUE
)
}
\arguments{
\item{inputdirectory}{path to folder containing datasets}

\item{outputdirectory}{path to folder where merged file is to be saved}

\item{outputname}{character string to name merged file}

\item{template}{logical (TRUE/FALSE) of whether the input data frame fits
the standard template}

\item{wide}{logical (TRUE/FALSE) of whether the output data frame should be
in wide format (default), else will produce long format data}
}
\description{
Merge weekly report monitoring and evaluation data.
Currently based on Guinea template; to be further in case of non-templated
documents
}
\author{
Alice Carr
}
