% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_M_and_E.R
\name{recode_kpi}
\alias{recode_kpi}
\title{Recode country monitoring and evaluation excel tool data to fit
new ODK based tool (key performance indicators)}
\usage{
recode_kpi(inputfile)
}
\arguments{
\item{inputfile}{a data frame produced by the merge_kpi function}
}
\description{
Recode country monitoring and evaluation excel tool data to fit
new ODK based tool (key performance indicators)
}
\author{
Alice Carr, Alex Spina
}
